requirements.txt
django
django-ipware
django-cors-headers

djangorestframework
djangorestframework_simplejwt

drf-yasg
djoser

pytest
pytest-benchmark
pytest-django
pytest-runner

python-decouple

ta
numpy
scipy
pandas
matplotlib

celery
redis

poloniexapi
easydict